Adjusted XDC file to reflect changed wbscope variable names
The problem manifests itself in the fact that Xilinx wants to know about
the clock crossing from the 25MHz ethernet clock to the 82MHz system clock.
These crossings are an error.  This XDC file tells Xilinx to ignore the
error.  The previous one did as well ... but the ignore information is
dependent upon knowing the starting and ending points of the logic, and
those names changed with the last update of wbscope.